+++ VanKornoDB/src/main/java/com/vankorno/vankornodb/get/ObjectsPro.kt
@@ -8,8 +8,7 @@ import com.vankorno.vankornodb.dbManagement.data.BaseEntity
 import com.vankorno.vankornodb.dbManagement.data.TableInfoBase
 
 /** 
- * Retrieves a list of objects mapped from the given columns. 
- * Similar to the reified version but uses explicit KClass parameter.
+ * Retrieves a list of objects mapped from the given columns.
  */
 fun <T : BaseEntity> SQLiteDatabase.getObjectsPro(                      tableInfo: TableInfoBase<T>,
                                                                               dsl: FullDsl.()->Unit,
@@ -31,11 +30,10 @@ fun <T : BaseEntity> SQLiteDatabase.getObjectsPro(                      tableInf
 
 /** 
  * Retrieves a map of objects from the given table.
- * Similar to the reified version but uses explicit KClass parameter. 
  */
 fun <T : BaseEntity> SQLiteDatabase.getObjMapPro(                       tableInfo: TableInfoBase<T>,
                                                                               dsl: FullDsl.()->Unit,
-): Map<Int, T> = getCursorPro(tableInfo.name) {
+): Map<Long, T> = getCursorPro(tableInfo.name) {
     applyDsl(dsl)
 }.use { cursor ->
     buildMap {
@@ -43,7 +41,7 @@ fun <T : BaseEntity> SQLiteDatabase.getObjMapPro(                       tableInf
             val idColIdx = cursor.getColumnIndexOrThrow("id")
             do {
                 val entity = cursor.toEntity(tableInfo.schema)
-                put(cursor.getInt(idColIdx), entity)
+                put(cursor.getLong(idColIdx), entity)
             } while (cursor.moveToNext())
         }
     }
